pyproject.toml
[tool.poetry]
name = "gpt-pyneapple"
version = "0.1.0"
description = "Integrate Pyneapple with GPT"
authors = ["Siddhant-0519"]
readme = "README.md"
packages = [{include = "gpt_pyneapple"}]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.25.0"
geopandas = "^0.13.2"
pandas = "^2.0.3"
libpysal = "^4.7.0"
matplotlib = "^3.7.1"
JPype1 = "^1.4.1"
openai = "^0.27.8"
fastapi = "^0.99.1"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
